+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Added texture of the Earth to directory
+- Applied texture on the planet
+- Made the planet rotate on itself
+
+## Changed
+- Made the planet an object with properties
 
 ## [0.0.8] - 2021-09-20
 ### Added